[TECH-404] Move env var extraction to a more logical namespace
branch: feature/TECH-404-link-up-deploy-set
